chore(plugin-report): remove dead useReportData aggregation pipeline
ADR-0021 moved report aggregation into the semantic layer (queryDataset),
so the pre-9.0 CLIENT-SIDE pipeline — the useReportData hook plus
buildAggregateQuery / groupAndAggregate / pivotRows / aggregateRows /
collectFields / columnKey / bucketDate / groupingValue and their result
types — had no remaining consumers (verified zero imports across objectui
and all five product repos). Delete the 840-line hook + its test.

The one symbol still in use, mergeFilters (the dataset renderer's
scope-filter combinator), moves to its own ./mergeFilters module and
stays exported from the package root. DatasetReportRenderer's import is
repointed; the barrel drops the dead exports.
